refactor: more import cleanup
