+++ packages/salesforcedx-vscode-apex/src/codecoverage/colorizer.ts
@@ -18,7 +18,7 @@ import { StatusBarToggle } from './statusBarToggle';
 
 const pathToApexTestResultsFolder = projectPaths.apexTestResultsFolder();
 
-export const getLineRange = (document: TextDocument, lineNumber: number): Range => {
+const getLineRange = (document: TextDocument, lineNumber: number): Range => {
   let adjustedLineNumber: number;
   let firstLine: TextLine;
   try {
