+++ packages/salesforcedx-vscode-core/src/commands/auth/orgLoginWebDevHub.ts
@@ -20,7 +20,7 @@ import { SfCommandlet, SfCommandletExecutor, SfWorkspaceChecker } from '../util'
 import { DEFAULT_ALIAS } from './authParamsGatherer';
 import { AuthDemoModeExecutor, OrgLoginWebContainerExecutor } from './orgLoginWeb';
 
-export class OrgLoginWebDevHubContainerExecutor extends OrgLoginWebContainerExecutor {
+class OrgLoginWebDevHubContainerExecutor extends OrgLoginWebContainerExecutor {
   public build(data: AuthDevHubParams): Command {
     const command = new SfCommandBuilder().withDescription(nls.localize('org_login_web_authorize_dev_hub_text'));
 
@@ -35,7 +35,7 @@ export class OrgLoginWebDevHubContainerExecutor extends OrgLoginWebContainerExec
   }
 }
 
-export class OrgLoginWebDevHubExecutor extends SfCommandletExecutor<{}> {
+class OrgLoginWebDevHubExecutor extends SfCommandletExecutor<{}> {
   protected showChannelOutput = false;
 
   public build(data: AuthDevHubParams): Command {
@@ -50,7 +50,7 @@ export class OrgLoginWebDevHubExecutor extends SfCommandletExecutor<{}> {
   }
 }
 
-export class OrgLoginWebDevHubDemoModeExecutor extends AuthDemoModeExecutor<{}> {
+class OrgLoginWebDevHubDemoModeExecutor extends AuthDemoModeExecutor<{}> {
   public build(data: AuthDevHubParams): Command {
     return new SfCommandBuilder()
       .withDescription(nls.localize('org_login_web_authorize_dev_hub_text'))
@@ -64,7 +64,7 @@ export class OrgLoginWebDevHubDemoModeExecutor extends AuthDemoModeExecutor<{}> 
   }
 }
 
-export class AuthDevHubParamsGatherer implements ParametersGatherer<AuthDevHubParams> {
+class AuthDevHubParamsGatherer implements ParametersGatherer<AuthDevHubParams> {
   public async gather(): Promise<CancelResponse | ContinueResponse<AuthDevHubParams>> {
     const aliasInputOptions = {
       prompt: nls.localize('parameter_gatherer_enter_alias_name'),
