+++ packages/salesforcedx-vscode-core/src/services/index.ts
@@ -4,5 +4,5 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-export { CoreLoggerService, getCoreLoggerService } from './getCoreLoggerService';
+export { getCoreLoggerService } from './getCoreLoggerService';
 export * from './telemetry';
