+++ packages/salesforcedx-vscode-core/src/services/telemetry/telemetryServiceProvider.ts
@@ -9,7 +9,7 @@ import { TelemetryService } from '@salesforce/salesforcedx-utils-vscode';
 import { TelemetryServiceInterface } from '@salesforce/vscode-service-provider';
 import * as vscode from 'vscode';
 
-export const getTelemetryServiceForKey = (key: string | undefined): Promise<TelemetryServiceInterface> => {
+const getTelemetryServiceForKey = (key: string | undefined): Promise<TelemetryServiceInterface> => {
   console.log(`key: ${key}`);
   return Promise.resolve(TelemetryService.getInstance(key));
 };
