--- packages/salesforcedx-vscode-lwc/src/testSupport/utils/index.ts
@@ -1,10 +1,0 @@
-/*
- * Copyright (c) 2019, salesforce.com, inc.
- * All rights reserved.
- * Licensed under the BSD 3-Clause license.
- * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
- */
-import { startWatchingEditorFocusChange } from './context';
-import { isLwcJestTest } from './isLwcJestTest';
-
-export { isLwcJestTest, startWatchingEditorFocusChange };
